requirements.txt
python-telegram-bot>=22.5
openai>=2.6.1
aiosqlite>=0.21.0
pyyaml>=6.0.3
python-dotenv>=1.2.1
aiofiles>=25.1.0
fastapi>=0.115.0
uvicorn>=0.34.0